Extract entity expansion into own class
